Add felony class A and person crime files
